Added .gitignore and some code cleanup

+++ File-Integrity-Scanner/src/main/java/org/pwss/msr/repository/ScanRepository.java
@@ -1,7 +1,7 @@
 package org.pwss.msr.repository;
 
-import org.pwss.msr.domain.model.entities.Scan;
+import org.pwss.msr.domain.model.entities.scan.Scan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScanRepository extends JpaRepository<Scan, Long> {
+public interface ScanRepository extends JpaRepository<Scan, Integer> {
 }
